feat: move endpoint tabs to app main page
