+++ apps/web/src/services/administration/components/AppAPIKeyRestEndpointsTab.vue
