+++ apps/web/src/services/administration/pages/AppMainPage.vue
@@ -2,7 +2,7 @@
 import { computed, onUnmounted, reactive } from 'vue';
 
 import {
-    PHorizontalLayout, PHeading, PButton,
+    PHorizontalLayout, PHeading, PButton, PTab,
 } from '@spaceone/design-system';
 import { cloneDeep } from 'lodash';
 
@@ -10,6 +10,8 @@ import { ROLE_TYPE } from '@/schema/identity/role/constant';
 import { store } from '@/store';
 import { i18n } from '@/translations';
 
+import AppAPIKeyGRPCEndpointsTab from '@/services/administration/components/AppAPIKeyGRPCEndpointsTab.vue';
+import AppAPIKeyRestEndpointsTab from '@/services/administration/components/AppAPIKeyRestEndpointsTab.vue';
 import AppManagementTable from '@/services/administration/components/AppManagementTable.vue';
 import { APP_DROPDOWN_MODAL_TYPE } from '@/services/administration/constants/app-constant';
 import { useAppPageStore } from '@/services/administration/store/app-page-store';
@@ -19,6 +21,18 @@ const appPageState = appPageStore.$state;
 
 const storeState = reactive({
     roleType: computed(() => store.state.user.roleType),
+});
+
+const tabs = [{
+    name: 'rest',
+    label: 'REST',
+}, {
+    name: 'gRPC',
+    label: 'gRPC',
+}];
+const state = reactive({
+    activeTab: 'rest',
+    userId: computed(() => store.state.user.userId),
 });
 
 /* Component */
@@ -59,6 +73,16 @@ onUnmounted(() => {
                 <app-management-table :table-height="height" />
             </template>
         </p-horizontal-layout>
+        <p-tab v-model="state.activeTab"
+               :tabs="tabs"
+        >
+            <template #rest>
+                <app-a-p-i-key-rest-endpoints-tab />
+            </template>
+            <template #gRPC>
+                <app-a-p-i-key-g-r-p-c-endpoints-tab />
+            </template>
+        </p-tab>
     </section>
 </template>
